Continuing to split upload and edge
