+++ CMakeLists.txt
@@ -53,6 +53,7 @@ add_subdirectory(src/emulator)
 add_subdirectory(src/infra)
 add_subdirectory(src/prof)
 add_subdirectory(src/proto)
+add_subdirectory(src/upload)
 add_subdirectory(src/util)
 add_subdirectory(src/wasm)
 add_subdirectory(src/worker)
