+++ ansible/pistache.yml
@@ -1,29 +1,32 @@
 ---
 
-- name: "Clone"
-  git:
-    repo: "https://github.com/oktal/pistache.git"
-    dest: "/usr/local/code/lib/pistache"
-    update: yes
+- hosts: localhost
+  gather_facts: no
+  tasks:
+    - name: "Clone"
+      git:
+        repo: "https://github.com/oktal/pistache.git"
+        dest: "/usr/local/code/lib/pistache"
+        update: yes
 
-- name: "Set up build"
-  shell: "{{ item }}"
-  args:
-    chdir: "/usr/local/code/lib/pistache"
-  with_items:
-  - "git submodule update --init"
-  - "mkdir -p build"
+    - name: "Set up build"
+      shell: "{{ item }}"
+      args:
+        chdir: "/usr/local/code/lib/pistache"
+      with_items:
+      - "git submodule update --init"
+      - "mkdir -p build"
 
-- name: "Build"
-  shell: "{{ item }}"
-  args:
-    chdir: "/usr/local/code/lib/pistache/build"
-  with_items:
-  - 'cmake -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Release ..'
-  - "make"
+    - name: "Build"
+      shell: "{{ item }}"
+      args:
+        chdir: "/usr/local/code/lib/pistache/build"
+      with_items:
+      - 'cmake -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Release ..'
+      - "make"
 
-- name: "Install"
-  become: yes
-  shell: "sudo make install"
-  args:
-    chdir: "/usr/local/code/lib/pistache/build"
+    - name: "Install"
+      become: yes
+      shell: "sudo make install"
+      args:
+        chdir: "/usr/local/code/lib/pistache/build"
